You can play the game again
But I think there might be an issue with huge memory leaks since I don't
know how to TRULY shut down one instance of a game and then restart it.
If we knew we wanted to do it this way, then we would've addressed that
from the main class but currently im too lazy to do that. onto the next
to dos!

I moved the accusation dialog parts into clueGame.
If a computer wins then a dialog will pop with the computer's name
If its a human win, simply replaced with "You"
but if a human makes a false accusation, only then will another dialog
pop up saying the accusation is false, no dialog pops up in the case of
an AI accusation